vertexai-mlops/experiments-sample/requirements.txt
# Vertex AI Experiments サンプル用依存関係

# Google Cloud
google-cloud-aiplatform>=1.38.0

# Kubeflow Pipelines
kfp>=2.4.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# データ処理
pandas>=2.0.0
numpy>=1.24.0

# 設定ファイル
pyyaml>=6.0
